Rename gregslist cleaner system and split hiring manager
