+++ src/gregslist/mod.rs
@@ -2,6 +2,6 @@ pub mod component;
 
 pub mod plugin;
 
-pub use plugin::GregslistPlugin;
+pub use plugin::{GregslistPlugin, gregslist_expiration_system};
 
 pub use component::{Advert, Gregslist, GregslistConfig, VacancyDirty};
