+++ src/hiring_manager/mod.rs
@@ -1,0 +1,3 @@
+pub mod plugin;
+
+pub use plugin::HiringManagerPlugin;
