+++ src/main.rs
@@ -6,6 +6,7 @@ use bevy::prelude::*;
 
 mod baby_spawner;
 mod gregslist;
+mod hiring_manager;
 mod inventory;
 mod jobs;
 mod mortality;
@@ -60,6 +61,7 @@ fn main() {
         .add_plugins(mortality::MortalityPlugin)
         .add_plugins(jobs::JobsPlugin)
         .add_plugins(gregslist::GregslistPlugin::new(60.0))
+        .add_plugins(hiring_manager::HiringManagerPlugin)
         .add_systems(Startup, spawn_jobs)
         //.add_systems(Startup, |mut time: ResMut<Time<Virtual>>| {
         //    time.set_relative_speed(DAY as f32);
